Attempt to simplify the syntax of the base HOAS
The main idea is that, to denote the return value of a spec,
we bind the value using a function `fun r : val => flow`, where
flow asserts facts about r.

For example, a spec for the value 1 is (fun r : val => ens \[r = vint 1]).
To assert that this spec holds for the value 1, we just "apply" the spec
with the value 1.

Thus, the syntax of ens is now just `hprop -> flow`, the result is binded
in the outer function. The nice thing about this representation is that,
the forward rule for let-binding is purely syntatic.