Finish proving unique decomposition.
The main insight is that: the proof of unique decomposition basically
contains an algorithm to "construct" the unique decomposition:

forall e, (exists v, e = v) \/ (exists! M R e', e = meta_rplug M (rplug R e'))

Notice the quantifier `exists`. In Rocq, we prove `exists` by giving the witness,
and we can only give the witness if we can construct it.

Therefore, a better idea is to extract the decomposition as a separate function,
which is much easier to study; and is useful to state our other lemmas as well.

After obtaining the decomposition function, we prove unique decomposition by
"delegating" to this decomposing function. We just need to prove that, given
a redex, an evaluation context and a meta context, plugging them and decomposing
them is identity. The prove is by pure equational reasoning (!) and it much easier
to understand.